requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.9
pyjwt==2.8.0
face-recognition==1.3.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.1
cryptography==41.0.4
scikit-learn==1.3.0
flasgger==0.9.5
pytest==7.4.0
pytest-flask==1.2.0
requests==2.31.0